Created Algebra.Graph.Relation.Symmetric.Internal for holding internal functions and definitions of SymmetricRelation.
Fixed comment/tests relation.
