+++ test/Main.hs
@@ -14,7 +14,6 @@ import Data.Graph.Test.Typed
 
 main :: IO ()
 main = do
-    testSymmetricRelation
     testAdjacencyIntMap
     testAdjacencyMap
     testExport
@@ -26,4 +25,5 @@ main = do
     testNonEmptyAdjacencyMap
     testNonEmptyGraph
     testRelation
+    testSymmetricRelation
     testTyped
